+++ src/components/episodeItem/episodeItem.css
